python/requirements.txt
psycopg2-binary
streamlit
openai
python-dotenv
streamlit_js_eval
better_profanity
cryptography
